migrate: move meta to components directory
- Migrated meta component files to src/components/meta/
- Moved and renamed index.tsx -> template.tsx
- Updated import statement in src/pages/[[...slug]].tsx from @/templates/common/meta to @/components/meta/template
- Removed empty src/templates/common/meta/ directory
- All TypeScript compilation checks passing
- No linter errors found

This meta component is a utility component that handles SEO meta tags including:
- Default Open Graph and Twitter Card tags
- Custom meta tags from resource.metatags
- Last updated timestamps
- iOS smart banner configuration for specific VA health and benefits pages
- Canonical URLs and robots meta tags
