+++ src/pages/[[...slug]].tsx
@@ -64,7 +64,7 @@ import HTMLComment from '@/templates/common/util/HTMLComment'
 import { Event } from '../components/event/template'
 import { EventListing } from '../components/eventListing/template'
 import { LocationsListing } from '../components/locationsListing/template'
-import { Meta } from '@/templates/common/meta'
+import { Meta } from '@/components/meta/template'
 import { NewsStory } from '../components/newsStory/template'
 import { PressRelease } from '../components/pressRelease/template'
 import { PressReleaseListing } from '../components/pressReleaseListing/template'
